Inject current Koin application from Android context
The Koin application is looked up in the Context tree and the Koin CompositionLocals are provided with/from this value.
This (still) requires this function to be called around the first time Koin is used from Compose.
A practical place for this is the setContent function.
This means that (boilerplate) code needs to be added to every Activity or Fragment using Compose, and every test case where a Composable is tested in isolation.

Untested - will need verification with these test frameworks!

Fixes gh-1557