Docs: Use `platform` for BOM in Version Catalog setup
Relates to #1721 

Failure to use `platform` for the BOM in a Kotlin Multiplatform project will cause iOS project failures during sync. This updates the setup docs so that the "Version Catalogs" setup is the same as the regular setup.

Additionally, replace `platform` with `project.dependencies.platform` due to the Kotlin deprecation: `'platform(Any): Dependency' is deprecated. Scheduled for removal in Kotlin 2.0`.