Fix a race condition when using `parametersOf`
`Scope` was incorrectly using `ThreadLocal` leading to a race
condition. The actual issue was not putting `ArrayDeque` in a
thread-local variable but creating the thread-local variable itself.
Several concurrent threads could have created several instances
of the `parameterStack` thread-local variable and only the last one
would have used.
The correct usage of `ThreadLocal` is to instantiate it directly
and let the `ThreadLocal` machinery to deal with synchronization.

This commit also adds a test that was failing prior to the fix.
I've also added `iosSimulatorArm64` to the `:core:koin-test-coroutines`
to test the fix on iOS.

This change fixes #2305.